error_bounds ready for float

+++ chain-demo/src/chain/mod.rs
@@ -38,9 +38,9 @@ pub type TxType = u64;
 // FloatType especially for linear regression
 pub type FloatType = f64;
 
-#[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Parameter {
-    pub error_bounds: u8,
+    pub error_bounds: FloatType,
     pub inter_index: bool,
     pub intra_index: bool,
     pub start_block_id: u64,
